Revert "Merge pull request #287 from Stanford-Online/azimpradhan/sysadmin-course-tabs"
This reverts commit 2e65165d2d40ae46e17d845b71e6c31674f22fbc, reversing
changes made to 92893bccc77b9269fb783010b10bbf3b109b8e87.
