+++ cms/djangoapps/contentstore/views/tabs.py
@@ -12,6 +12,7 @@ from django.views.decorators.http import require_http_methods
 
 from edxmako.shortcuts import render_to_response
 from xmodule.modulestore.django import modulestore
+from xmodule.modulestore import ModuleStoreEnum
 from xmodule.tabs import CourseTabList, CourseTab, InvalidTabsException, StaticTab
 from opaque_keys.edx.keys import CourseKey, UsageKey
 
@@ -167,3 +168,36 @@ def get_tab_by_locator(tab_list, usage_key_string):
         url_slug=item.location.name,
     )
     return CourseTabList.get_tab_by_id(tab_list, static_tab.tab_id)
+
+
+# "primitive" tab edit functions driven by the command line.
+# These should be replaced/deleted by a more capable GUI someday.
+# Note that the command line UI identifies the tabs with 1-based
+# indexing, but this implementation code is standard 0-based.
+
+def validate_args(num, tab_type):
+    "Throws for the disallowed cases."
+    if num <= 1:
+        raise ValueError('Tabs 1 and 2 cannot be edited')
+    if tab_type == 'static_tab':
+        raise ValueError('Tabs of type static_tab cannot be edited here (use Studio)')
+
+
+def primitive_delete(course, num):
+    "Deletes the given tab number (0 based)."
+    tabs = course.tabs
+    validate_args(num, tabs[num].get('type', ''))
+    del tabs[num]
+    # Note for future implementations: if you delete a static_tab, then Chris Dodge
+    # points out that there's other stuff to delete beyond this element.
+    # This code happens to not delete static_tab so it doesn't come up.
+    modulestore().update_item(course, ModuleStoreEnum.UserID.primitive_command)
+
+
+def primitive_insert(course, num, tab_type, name):
+    "Inserts a new tab at the given number (0 based)."
+    validate_args(num, tab_type)
+    new_tab = CourseTab.from_json({u'type': unicode(tab_type), u'name': unicode(name)})
+    tabs = course.tabs
+    tabs.insert(num, new_tab)
+    modulestore().update_item(course, ModuleStoreEnum.UserID.primitive_command)
