+++ lms/djangoapps/dashboard/sysadmin_urls.py
@@ -10,7 +10,6 @@ urlpatterns = patterns(
     '',
     url(r'^$', sysadmin.Users.as_view(), name="sysadmin"),
     url(r'^courses/?$', sysadmin.Courses.as_view(), name="sysadmin_courses"),
-    url(r'^course_tabs/?$', sysadmin.CourseTabs.as_view(), name="sysadmin_course_tabs"),
     url(r'^staffing/?$', sysadmin.Staffing.as_view(), name="sysadmin_staffing"),
     url(r'^gitlogs/?$', sysadmin.GitLogs.as_view(), name="gitlogs"),
     url(r'^gitlogs/(?P<course_id>.+)$', sysadmin.GitLogs.as_view(),
