+++ lms/static/sass/multicourse/_sysadmin.scss
@@ -49,11 +49,3 @@
 .mgmt-commands-body .checkbox-label {
     display: inline;
 }
-
-.edit-course-tabs .long-input {
-  width: 60%;
-}
-
-.edit-course-tabs .unique-row {
-  margin-top: 20px;
-}
