+++ lms/templates/sysadmin_dashboard.html
@@ -7,14 +7,10 @@ from django.utils.translation import ugettext as _
 
 <%block name="headextra">
   <%static:css group='style-course'/>
-  <link rel="stylesheet" type="text/css" href="${static.url('css/sysadmin.css')}">
   <script type="text/javascript" src="${static.url('js/vendor/flot/jquery.flot.js')}"></script>
   <script type="text/javascript" src="${static.url('js/vendor/flot/jquery.flot.axislabels.js')}"></script>
   %if modeflag.get('mgmt_commands'):
       <script type='text/javascript' src="${static.url('js/sysadmin/mgmt_commands.js')}"></script>
-  %endif
-  %if modeflag.get('course_tabs'):
-    <script type="text/javascript" src="${static.url('js/sysadmin/course_tabs.js')}"></script>
   %endif
 </%block>
 
@@ -64,12 +60,11 @@ textarea {
     <h2 class="instructor-nav">
 	  <a href="${reverse('sysadmin')}" class="${modeflag.get('users')}">${_('Users')}</a>
       <a href="${reverse('sysadmin_courses')}" class="${modeflag.get('courses')}">${_('Courses')}</a>
-      <a href="${reverse('sysadmin_course_tabs')}" class="${modeflag.get('course_tabs')}">${_('Course Tabs')}</a>
-      <a href="${reverse('sysadmin_staffing')}" class="${modeflag.get('staffing')}">${_('Staffing and Enrollment')}</a>
+      <a href="${reverse('sysadmin_staffing')}" class="${modeflag.get('staffing')}">${_('Staffing and Enrollment')}</a> 
       ## Translators: refers to http://git-scm.com/docs/git-log
       <a href="${reverse('gitlogs')}">${_('Git Logs')}</a>
       <a href="${reverse('sysadmin_task_queue')}" class="${modeflag.get('task_queue')}">${_('Task Queue')}</a>
-      <a href="${reverse('sysadmin_mgmt_commands')}" class="${modeflag.get('mgmt_commands')}">${_('CMDS')}</a>
+      <a href="${reverse('sysadmin_mgmt_commands')}" class="${modeflag.get('mgmt_commands')}">${_('MGMT Commands')}</a>
     </h2>
 	<hr />
 %if modeflag.get('users'):
@@ -149,41 +144,10 @@ textarea {
 
 %endif
 
-%if modeflag.get('course_tabs'):
-<form class="sysadmin-course-tabs-form" name="action" method="POST">
-  <input type="hidden" name="csrfmiddlewaretoken" value="${ csrf_token }" />
-  <div class="edit-course-tabs">
-      <h4>Edit Course Tabs</h4>
-      <ul class="list-input">
-          <li class="field text">
-              <label for="course_id">
-                  ${_('Course ID')}:
-              </label>
-              <input type="text" name="course_id" class="long-input"/>
-          </li>
-      </ul>
-      <div class="form-actions unique-row">
-        <button type="submit" name="action" value="get_current_tabs">${_('Get Current Tabs')}</button>
-      </div>
-      <div class="form-actions unique-row">
-        <button class="delete-tab-button" type="button" value="delete_tab">${_('Delete Tab')}</button>
-        <input type="text" name="tab_delete"/>
-        <span>(sample argument: 7)</span>
-      </div>
-      <div class="form-actions unique-row">
-        <button class="insert-tab-button" type="button" value="insert_tab">${_('Insert Tab')}</button>
-        <input type="text" name="tab_insert"/>
-        <span>(sample arguments: 7, static_tab, Chat)</span>
-      </div>
-  </div>
-</form>
-<hr style="width:40%" />
-%endif
-
 %if modeflag.get('courses'):
 <h3>${_('Administer Courses')}</h3><br/>
 
-<form class="sysadmin-course-form" name="action" method="POST">
+<form name="action" method="POST">
   <input type="hidden" name="csrfmiddlewaretoken" value="${ csrf_token }" />
   <ul class="list-input">
 	<li class="field text">
@@ -248,9 +212,7 @@ textarea {
 %endif
 
 %if msg:
-    <div class="server-message">
-        <p>${msg}</p>
-    </div>
+    <p>${msg}</p>
 %endif
 
 %if modeflag.get('mgmt_commands'):
