+++ lms/templates/sysadmin_dashboard_gitlogs.html
@@ -114,12 +114,11 @@ textarea {
       <h2 class="instructor-nav">
         <a href="${reverse('sysadmin')}">${_('Users')}</a>
         <a href="${reverse('sysadmin_courses')}">${_('Courses')}</a>
-        <a href="${reverse('sysadmin_course_tabs')}">${_('Course Tabs')}</a>
         <a href="${reverse('sysadmin_staffing')}">${_('Staffing and Enrollment')}</a>
         ## Translators: refers to http://git-scm.com/docs/git-log
         <a href="${reverse('gitlogs')}" class="active-section">${_('Git Logs')}</a>
         <a href="${reverse('sysadmin_task_queue')}">${_('Task Queue')}</a>
-        <a href="${reverse('sysadmin_mgmt_commands')}">${_('CMDS')}</a>
+        <a href="${reverse('sysadmin_mgmt_commands')}">${_('MGMT Commands')}</a>
       </h2>
       <hr />
